Removed Unused Properties
- Removed Unused Properties
  - Call.ParentCall
  - ChartView.Horizontal
  - LiveChartSeries.UseDateTimeAxis
  - TimeRangePeriod.Name
  - TimeRangeValue.Name
  - TabChart.TabInstance and TabInstance.IdxListToListSeries
  - Serializer.TaskInstance

+++ Tests/SideScroll.Serialize.Tests/LargeFileTests.cs
@@ -141,10 +141,10 @@ public struct MyColor
 	// Make the view brighter
 	public void Brighten(short value)
 	{
-		Red = (short)Math.Min(short.MaxValue, (int)Red + value);
-		Green = (short)Math.Min(short.MaxValue, (int)Green + value);
-		Blue = (short)Math.Min(short.MaxValue, (int)Blue + value);
-		Alpha = (short)Math.Min(short.MaxValue, (int)Alpha + value);
+		Red = (short)Math.Min(short.MaxValue, Red + value);
+		Green = (short)Math.Min(short.MaxValue, Green + value);
+		Blue = (short)Math.Min(short.MaxValue, Blue + value);
+		Alpha = (short)Math.Min(short.MaxValue, Alpha + value);
 	}
 }
